Changes to app folder
I moves the stock.service.ts file to the app folder for consistency. Also fixed misplaced closing bracket in stock.service file. Removed properties from auth.service.ts file.
